# Improve initial setup time and memory consumption in fast histogram