This is a response to issue #2326.
* Modified implementation for approximate quantile sketch for reduced memory consumption
* Parallel construction of columnar access structure -- this is a transposed version of input matrix to speed up column-wise access.
* Option (`use_columnar_access=0`) to disable columnar access structure entirely, to further reduce initial setup time and memory usage. Column access may slow down with this option.
* Slightly improved feature grouping algorithm