@hcho3 2017-07-24
Here are some results on a c3.8xlarge instance (60 GB RAM, 32 cores):

Configuration | setup time | peak memory usage | full log
--------------|-------------|-----------|--------
`enable_feature_grouping=1, use_columnar_access=1` | 190 sec | 30.5 GB | [log.txt](https://github.com/dmlc/xgboost/files/1170864/log.txt)
`enable_feature_grouping=0, use_columnar_access=1`| 69 sec | 12.4 GB | [log2.txt](https://github.com/dmlc/xgboost/files/1170865/log2.txt)
`enable_feature_grouping=0, use_columnar_access=0` | 40 sec | 12.4 GB | [log3.txt](https://github.com/dmlc/xgboost/files/1170866/log3.txt)