@RAMitchell 2017-07-25
Is there any reason we can't use a more conventional weighted quantile algorithm (i.e. sort based) here instead of a sketch?