@tqchen 2017-07-28
@hcho3 let me know if it is ready to merge