@Laurae2 2017-08-01
@hcho3 It crashes on my custom reputation dataset (2,250,000 observations x 23,636 features) when it reaches the grouping feature part. If I disable feature grouping, it works (after 18 minutes of columnar access generation).

![image](https://user-images.githubusercontent.com/9083669/28823363-784ba8d0-76bd-11e7-883d-675172c75267.png)