@hcho3 2017-08-01
@Laurae2 Sorry for the late reply. The improvement was sorely concerned with improving setup 
time. Thanks again for your contribution. I do have a few questions for you.

1) When you say "custom reputation dataset," are you referring to the same dataset you used in issue #2326? I had run tests using the dataset given in #2326 with feature grouping enabled, without any crashes. Let me take another look into this.

2) Are you still using this machine?
* Quad E7-88xx v4 (72 cores), but 3 CPUs were disabled to get rid of NUMA
* 1TB RAM 1866MHz RAM (256GB RAM available to get rid of NUMA)
* Not virtualized
I'm trying to guess whether feature grouping crashes due to lack of memory.

3) Do you have the full log for the Bosch run (with `debug_verbose=1`)?

@tqchen I do need to take another look at this pull request. While I'm at it, I'm also inclined to completely re-write the feature grouping logic to make it more parallel. (The logic is inherently sequential now, as it has to inspect one feature at a time.) Let me get back to you on this.