@Laurae2 2017-08-01
@hcho3 Issue is probably here: https://github.com/hcho3/xgboost/blob/14a33f6cdf2b388f64293b0ac0718d5ab945b37f/src/common/column_matrix.h#L157-L189 (threads are probably sharing a common variable, leading to negative scalability)

1. It is not exactly the same dataset. My custom reputation dataset can be found using this preprocessing, to do in order:

* https://github.com/Laurae2/gbt_benchmarks/blob/master/new_core/creator/reput_setup1.R
* https://github.com/Laurae2/gbt_benchmarks/blob/master/new_core/creator/reput_setup2.R
* https://github.com/Laurae2/gbt_benchmarks/blob/master/new_core/creator/reput_setup3.R

It leads to the creation of a very large RDS file which contains the dataset (see https://github.com/dmlc/xgboost/issues/2326#issuecomment-313915978 for the training script).

2. Yes, still using that machine, but with all cores/RAM. I had over 950GB free RAM before it crashed. I noticed the CPU usage was around 5%, xgboost was trying to leverage all cores but it couldn't do so efficiently. Disabling 

3. Unfortunately my server is not available, therefore I tried to reproduce the log on my laptop, It seems I cannot reproduce it on the laptop (I had RAM swap at feature grouping, therefore it is normal it is slow at that part).

However, when it took 8min30 on Bosch I had the same issue as 2., e.g low CPU usage. The very slow part was when xgboost was trying to use all available cores. More exactly, I think it was here: https://github.com/hcho3/xgboost/blob/14a33f6cdf2b388f64293b0ac0718d5ab945b37f/src/common/column_matrix.h#L157-L189

Log for my laptop, i7-4600U:

```r
> model <- xgb.train(params = list(nthread = 4,
+                                  max_depth = 6,
+                                  num_leaves = 63,
+                                  tree_method = "hist",
+                                  grow_policy = "depthwise",
+                                  eta = 0.05,
+                                  max_bin = 255,
+                                  eval_metric = "auc",
+                                  debug_verbose = 2,
+                                  enable_feature_grouping = 1),
+                    data = train,
+                    nrounds = 1,
+                    watchlist = list(test = test),
+                    verbose = 2,
+                    early_stopping_rounds = 50)
[21:05:28] Tree method is selected to be 'hist', which uses a single updater grow_fast_histmaker.
[21:05:29] amalgamation/../src/common/hist_util.cc:37: Generating sketches...
[21:05:33] amalgamation/../src/common/hist_util.cc:75: Computing quantiles for features [0, 969)...
[21:05:33] amalgamation/../src/tree/updater_fast_hist.cc:70: Quantizing data matrix entries into quantile indices...
[21:05:37] amalgamation/../src/tree/updater_fast_hist.cc:75: Generating columnar access structure...
[21:05:49] amalgamation/../src/tree/updater_fast_hist.cc:82: Grouping features together...
[21:08:17] amalgamation/../src/tree/updater_fast_hist.cc:92: Done initializing training: 167.956 sec
[21:08:24] amalgamation/../src/tree/updater_prune.cc:74: tree pruning end, 1 roots, 82 extra nodes, 0 pruned nodes, max_depth=6
[21:08:24] amalgamation/../src/tree/updater_fast_hist.cc:273: 
InitData:          0.0625 ( 0.83%)
InitNewNode:       0.0000 ( 0.00%)
BuildHist:         7.2841 (96.83%)
EvaluateSplit:     0.0499 ( 0.66%)
ApplySplit:        0.1260 ( 1.67%)
========================================
Total:             7.5224
[21:08:29] amalgamation/../src/gbm/gbtree.cc:274: CommitModel(): 4.39619 sec
[21:08:29] amalgamation/../src/learner.cc:373: EvalOneIter(): 0.0469153 sec
[1]	test-auc:0.606967 
Will train until test_auc hasn't improved in 50 rounds.

[21:08:32] amalgamation/../src/tree/updater_prune.cc:74: tree pruning end, 1 roots, 82 extra nodes, 0 pruned nodes, max_depth=6
```