@khotilov 2017-08-02
@hcho3 parallelizing this part would be useful. I've also noticed that during the initialization lots of time was spent with just a single thread working.