@hcho3 2017-08-02
I had accidentally closed the pull request. Sorry for any confusion caused.